backend/requirements.txt
fastapi>=0.116.0
uvicorn[standard]>=0.35.0
mangum>=0.19.0
pydantic>=2.11.0
python-dotenv>=1.1.0
pymongo>=4.9.0
motor>=3.6.0
httpx>=0.28.0
